Prabin roles and permission

+++ app/Http/Controllers/Admin/CrewProfilesController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\CrewProfile;
 use App\Models\User;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\DB;
 
@@ -13,18 +14,21 @@ class CrewProfilesController extends Controller
 {
     public function index()
     {
+        Gate::authorize('index_crew_profiles');
         $crew = CrewProfile::with('user')->latest()->get();
         return view('layouts.admin.crew_profiles.index', compact('crew'));
     }
 
     public function create()
     {
+        Gate::authorize('create_crew_profiles');
         $users = User::all();
         return view('layouts.admin.crew_profiles.create', compact('users'));
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('create_crew_profiles');
         $request->validate([
             'role' => 'required|in:driver,helper',
             'license_number' => 'nullable|string',
@@ -73,6 +77,8 @@ class CrewProfilesController extends Controller
 
     public function edit(CrewProfile $crew_profile)
     {
+        Gate::authorize('update_crew_profiles');
+        
         $users = CrewProfile::select('users.name as crew_member_name','crew_profiles.*')
         ->join('users','users.id','=','crew_profiles.user_id')
         ->get();
@@ -89,6 +95,7 @@ class CrewProfilesController extends Controller
 
     public function update(Request $request, CrewProfile $crew_profile)
     {
+        Gate::authorize('update_crew_profiles');
         $request->validate([
             'role' => 'required|in:driver,helper',
             'license_number' => 'nullable|string',
@@ -135,12 +142,14 @@ class CrewProfilesController extends Controller
 
     public function show(CrewProfile $crew_profile)
     {
+        Gate::authorize('read_crew_profiles');
         $crew_profile->load('user');
         return view('layouts.admin.crew_profiles.show', compact('crew_profile'));
     }
 
     public function destroy(CrewProfile $crew_profile)
     {
+        Gate::authorize('delete_crew_profiles');
         $crew_profile->delete();
 
         return redirect()->route('admin.crew_profiles.index')
